Include coordinates mixin in PlaidPlot component.
This allows us to set sensible defaults for the PlaidPlot component,
that can be overwritten with either a user-supplied plotArea object, or
via the area helper.

Also, removes the yOffset of 8 from the yOffset, as it is a default that
leads to confusion.

Something we found in the changes meant for this PR, is the complexity
in trying to automatically set axis x and y positions. The fundamental
problem with it, is that when creating the visualization, maximum-plaid
doesn't know of the graph's orientation. In the process, it doesn't know
whether to prioritize the top or bottom margins for setting the starting
draw position of the axes(and the graph itself). We should think about
adding an 'orientation' system to the entire plaid-plot
component.[Issue](https://github.com/ivanvanderbyl/maximum-plaid/issues/14)

Thanks for help in developing this to @audsull.